Simplify box usage. Part 1/2: Vendor UI
- Yield only one page for a box,
- add representative item barcode to that page.

Does not restrict how the boxes are handled in future: checkout can have
either a list of box-item barcodes, a list of box-number barcodes, or
nothing of those to actually handle selling box items.
Old statistics etc work as is, but special handling of the boxes'
representative item barcode needs to be implemented later.