Implement dense normal distribution for general statistics.
Dygraph seems to limit graph to have only one x-axis. This greatly
reduces the detail of normal distribution graph that is provided only
for reference and does not follow the bucket graph data density.
Paint the normal distribution separately under the graph and use some of
its values in the data provided to Dygraph to have the distribution data
to be accurately represented.