Bind Counters to 1:1 clients.
Previously same Counter instance could have been used on multiple
clients simultaneously. While probably not a problem, it was not
intended.
Add a private key that is used instead of the counter identifier when
re-validating counter or logging clerk in so that if the key is changed
in the server, the previous client is rendered unusable.