Add separate setting to disable registration
The registration end time should always be available for UITexts, but
without that actually opening the registration.
Instead of using the end time field for two purposes, split it to two.
The migration will set all existing Events to have disabled=false, to
mirror old behavior. For new Events the setting defaults to true.